Add a very simple lexer generator.
The lexemes must be completely disjoint, such that by observing a
single character we can immediately decide which terminal (if any) it
belongs to.

Whitespace is hardcoded.

This obviously needs significantly more work, but it's actually enough
to parse Lisp expressions such as

    (define (fib n) (if (= n 0) 1 (mul n (fib (dec 1)))))

which with the sexp.cg grammar gives

    [1, 3, 0, 3, 1, 3, 0, 3, 0, 2, 3, 1, 3, 0, 3, 1, 3, 0, 2, 3, 1, 3, 0, 3, 0, 3, 1, 3, 0, 3, 1, 3, 0, 2, 2, 2, 2, 2]